supreme-commander-agent/requirements.txt
# Production dependencies
# Generated from pyproject.toml

google-cloud-aiplatform>=1.60.0
google-adk-python>=0.1.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
firebase-admin>=6.5.0
vertexai>=1.60.0
google-cloud-discoveryengine>=0.11.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-gcp-trace>=1.6.0
pydantic>=2.8.0
python-dotenv>=1.0.0
